synthesis/pyproject.toml
[project]
name = "ingestion-service"
version = "0.1.0"
description = "Ingestion and synthesis microservice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.30",
  "alembic>=1.14",
  "pgvector>=0.3",
  "aioboto3>=13",
  "anthropic>=0.40",
  "httpx>=0.28",
  "numpy>=2.1",
  "scipy>=1.14",
  "python-multipart>=0.0.12",
  "structlog>=24.4",
  "pyjwt>=2.9",
  "passlib[bcrypt]>=1.7",
  "bcrypt>=4.0,<5",
  "voyageai>=0.3",
  "openai>=1.55",
  "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.8",
  "mypy>=1.13",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ASYNC"]

